Implemented matcher between userId and detectorId
