+++ imports/startup/server/register-api.js
@@ -10,6 +10,7 @@ import '../../api/users/methods.js';
 import '../../api/locations/server/publications.js';
 import '../../api/locations/methods.js';
 import '../../api/detectors/server/publications.js';
+import '../../api/detectors/methods.js';
 import '../../api/coordinator/methods.js';
 import '../../api/submissions/methods.js';
 import '../../api/submissions/server/publications.js';
